In the 'PrimarySideBar' component, the color styles for buttons were moved out of the 'PrimaryLinkAsButton' component because these styles were not relevant to it.
In the 'BlockRecommendations' component, the link that previously directed to the recommendations page was replaced with a 'div' element. This 'div' element now receives a 'props' function to enable sorting of movies by recommendation. Additionally, within the 'map' method that creates an array of movie cards, each card now has a unique 'key' attribute assigned to it. Furthermore, the movie sorting logic was implemented in the 'SidebarWithRecommendation' component.