In the 'Header.css' file, the styling was modified by removing the 'padding' property and adjusting the 'gap'.
In the 'Logo.css' file, a new style was added to prevent the logo from being highlighted when the user is on the home page.

In the 'SortingPanel' component, the class names were changed to be more unique in order to avoid conflicts with classes used in other sections. Additionally, the 'margin' of the parent block was adjusted.

In the 'App.css' file, the height of the page was corrected.

In the 'AppColors.css' file, new colors were introduced for the 'Page navigation buttons'.

In the 'BigMovieCard' component, a more distinctive class name was assigned to the element that contains the information about the movie's release year.