The 'Error' component has undergone refactoring to improve code readability and maintainability. It has been divided into smaller, more manageable components for better organization.
The 'SearchForm' component now includes user input validation before sending the 'http' request. The 'sortByTitle' function has been modified accordingly to accept a validated value for the request and a separate value to display information about the user's input.

Functionality has been added to the 'Footer' and 'Logo' components to enable page reloading, allowing users to easily refresh the home page.

The 'isCarsLoading' state has been removed from the 'BlockRecommendations' component. This change was necessary to address issues with element behavior during loading and to ensure proper re-rendering of loading cards when the page is refreshed with updated data.

The styles that were previously defined in the 'Logo.css', 'Footer.css', and 'Homepage.css' files have been removed. These changes are made to give the 'Logo' and 'Footer' components back a visual indication that they are active.