The code underwent several significant changes. Functions like 'getAllMovies', 'getTopMovies', 'getRecommendedMovies', and 'getAllActors' now fetch data from the 'Airtable' service rather than a local server. This transition to the 'Airtable API' brought about some limitations, prompting adjustments in pagination. The 'MainNavButtons' component, once responsible for direct page navigation, has been removed. The 'PageNavigationButtons' component now exclusively offers options to move to the next, previous, and first page. Additionally, a new state called 'offsetPages' has been integrated into the 'AppStateProvider'. This state manages the 'offset' property, which directs to the next batch of data from the 'Airtable' table. By manipulating this data and incorporating it into subsequent requests, we facilitate navigation through pages. Omitting this parameter allows us to return to the initial page.
Given that we now rely on the 'Airtable API' for data retrieval, the sorting parameters (searchParams) in the 'AppStateProvider' have been updated to align with the specifications of the 'Airtable API'. Moreover, the 'searchParams' state is managed and updated using 'useReducer' instead of 'useState' for more efficient handling. In light of these alterations in pagination, the 'Pages.js' utility, previously used to calculate the number of movie pages and generate an array of page numbers, has been deprecated due to redundancy.

Furthermore, due to variations in data structure obtained from the 'Airtable API', adjustments have been made to the hooks 'useFetchingActors', 'useFetchingMainMovies', and 'useFetchingMovies'. These changes enable the transformation of resulting data into a format compatible with the application.