All sorting, data retrieval, and transmission on the 'Navigator' and 'MoviePage' pages have been migrated to the 'Airtable API'. In 'MoviesService', several adjustments were made: the function 'getMovieById' is now handled by the 'Airtable API', and 'updateMovieData' was renamed to 'rateMovie' for improved clarity, and this function was also moved to the 'Airtable API'.
Due to alterations in the 'Airtable' database structure where the 'id' field no longer holds the movie's serial number but rather the 'Airtable record id', all sorting functions that formerly utilized this field have now been adapted to use the 'serialNumber' field, which presently holds the serial number values of the movies.

Additionally, a new variable has been introduced in 'DataToSort' that contains the 'id' array of all the movies. This array is now employed in a function to retrieve a random movie. Finally, the 'Sorting' utility was removed since it is no longer utilized anywhere.