A set of new components, including 'Select', 'Options', and 'LoaderForOptions', has been introduced. Together, they emulate the standard HTML element '<select>' with some enhancements. These include a different HTML structure, the capability to load a fresh set of selection options along with a loading indicator, and consistent display across all devices and browsers.
Given the constraints of the 'Airtable API', an additional list of actors needs to be loaded. To accommodate this, several adjustments have been made. The 'getAllActors' function now accepts an 'offset' as an argument. A new state has been added to 'AppStateProvider' to store the 'offset' for loading the subsequent batch of actors. The 'useFetchingActors' hook now takes 'offset' and 'setOffset' as arguments, and it updates the 'offsetActors' value in the 'AppStateProvider' on each load. As a result of these changes, the 'Select' component has replaced 'Filter', 'FilterOptions', and 'Selector', with the latter components being subsequently removed.

Additionally, modifications have been implemented in the hooks 'useSortByActor', 'useSortByCountry', 'useSortByDirector', 'useSortByGenre', and 'useSortByYear'. These adjustments address an issue with the 'filterByFormula' field. Since the strings being sent may contain single quotes, and the 'Airtable API' requires wrapping values for comparison in quotes (single or double), we now utilize double quotes to prevent conflicts with single quotes on the 'Airtable' side. Furthermore, for compatibility with the new 'Select' component, the 'useSortByYear' hook now expects a string as an argument. The 'generateYearsArray' function has been updated to generate an array of strings instead of numbers.

Finally, an error occurring during the loading of the actors' list in the 'useHomepage' and 'useNavigator' hooks is now ignored. This ensures that non-essential data loading does not disrupt the user interface. Additionally, data on the 'Hompage' page is reloaded only when it is absent.